+++ packages/editor/editor-core/src/__tests__/integration/_helpers.ts
@@ -22,7 +22,7 @@ export const insertMention = async (browser: any, query: string) => {
   await browser.type(editable, '@');
   await browser.waitForSelector(typeAheadPicker);
   await browser.type(editable, query);
-  await browser.type(editable, 'Return');
+  await browser.keys(['Return']);
 };
 
 export const gotoEditor = async (browser: any) => {
@@ -182,7 +182,7 @@ export const insertMedia = async (
   const mediaCardCount = get$$Length(existingMediaCards) + filenames.length;
 
   // Workaround - we need to use different wait methods depending on where we are running.
-  if (browser.browser.desiredCapabilities) {
+  if (browser.browser.capabilities) {
     await browser.browser.waitUntil(async () => {
       const mediaCards = await browser.$$(mediaCardSelector);
 
@@ -195,7 +195,7 @@ export const insertMedia = async (
       );
     });
   } else {
-    await browser.evaluate(() => {
+    browser.evaluate(() => {
       window.scrollBy(0, window.innerHeight);
     });
     await browser.waitFor(
@@ -306,7 +306,7 @@ export const insertMenuItem = async (browser: any, title: string) => {
 };
 
 export const currentSelectedEmoji = '.emoji-typeahead-selected';
-export const typeahead = '.ak-emoji-typeahead';
+export const typeahead = '.ak-emoji-typeahead-list';
 
 export const insertEmoji = async (browser: any, query: string) => {
   await browser.type(editable, ':');
@@ -324,7 +324,7 @@ export const insertEmojiBySelect = async (browser: any, select: string) => {
 };
 
 export const currentSelectedEmojiShortName = async (browser: any) => {
-  return await browser.$(currentSelectedEmoji).getProperty('data-emoji-id');
+  return await browser.getProperty(currentSelectedEmoji, 'data-emoji-id');
 };
 
 export const highlightEmojiInTypeahead = async (
@@ -337,12 +337,12 @@ export const highlightEmojiInTypeahead = async (
     if (selectedEmojiShortName === `:${emojiShortName}:`) {
       break;
     }
-    await browser.type(editable, 'ArrowDown');
+    await browser.keys(['ArrowDown']);
   }
 };
 
 export const emojiItem = (emojiShortName: string): string => {
-  return `span[data-emoji-short-name=":${emojiShortName}:"]`;
+  return `span[shortname=":${emojiShortName}:"]`;
 };
 
 interface ResizeOptions {