+++ packages/media/media-ui/src/InlineCard/IconAndTitleLayout/styled.ts
@@ -1,19 +1,40 @@
 import styled from 'styled-components';
 import { HTMLAttributes, ComponentClass } from 'react';
 
-// TODO: Replace overrides with proper AtlasKit solution.
-export const IconWrapper: ComponentClass<HTMLAttributes<{}>> = styled.span`
-  line-height: 12px;
-  margin-right: 4px;
-  vertical-align: text-bottom;
-  user-select: none;
+// TODO: remove this override behaviour for @atlaskit/icon-object
+export const IconObjectOverrides = `
   & > span {
     height: 14px;
     width: 14px;
+    vertical-align: text-bottom;
     & > svg {
       vertical-align: text-top;
     }
   }
+`;
+// TODO: remove this override behaviour for @atlaskit/icon
+export const IconOverrides = `
+  & > span > span {
+    height: 14px;
+    width: 14px;
+    vertical-align: text-bottom;
+    & > svg {
+      vertical-align: text-top;
+    }
+  }
+`;
+
+// Wraps all icons represented in Inline Links. Icons have three sources/types:
+// - JSON-LD: from the generator.icon property coming back from ORS.
+// - @atlaskit/icon: for lock icons, unauthorized, etc.
+// - @atlaskit/icon-object: for object icons, e.g. repository, branch, etc.
+// NB: the first set of overrides style icons imported from @atlaskit/icon-object correctly.
+// NB: the second set of overrides style icons imported from @atlaskit/icon correctly.
+export const IconWrapper: ComponentClass<HTMLAttributes<{}>> = styled.span`
+  margin-right: 4px;
+  user-select: none;
+  ${IconOverrides}
+  ${IconObjectOverrides}
 `;
 
 // The main 'wrapping' element, title of the content.