+++ packages/search/global-search/src/__tests__/unit/confluence/ConfluenceQuickSearchContainer.test.tsx
@@ -36,6 +36,7 @@ function render(partialProps?: Partial<Props>) {
     peopleSearchClient: noResultsPeopleSearchClient,
     useQuickNavForPeopleResults: false,
     useCPUSForPeopleResults: false,
+    fasterSearchFFEnabled: false,
     logger,
     ...partialProps,
   };
@@ -87,6 +88,33 @@ describe('ConfluenceQuickSearchContainer', () => {
     });
   });
 
+  it('should call cross product search client with correct query version', async () => {
+    const searchSpy = jest.spyOn(noResultsCrossProductSearchClient, 'search');
+    const dummyQueryVersion = 123;
+
+    const wrapper = render({
+      confluenceClient: noResultsConfluenceClient,
+      crossProductSearchClient: noResultsCrossProductSearchClient,
+    });
+
+    const quickSearchContainer = wrapper.find(QuickSearchContainer);
+    (quickSearchContainer.props() as QuickSearchContainerProps).getSearchResults(
+      'query',
+      sessionId,
+      100,
+      dummyQueryVersion,
+    );
+
+    expect(searchSpy).toHaveBeenCalledWith(
+      'query',
+      expect.any(Object),
+      expect.any(Array),
+      dummyQueryVersion,
+    );
+
+    searchSpy.mockRestore();
+  });
+
   it('should return ab test data', async () => {
     const abTest: ABTest = {
       abTestId: 'abTestId',
@@ -130,6 +158,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults).toMatchObject({
@@ -156,7 +185,7 @@ describe('ConfluenceQuickSearchContainer', () => {
     });
   });
 
-  it('should use CPUs for people results when enabled', async () => {
+  it('should use CPUS for people results when enabled', async () => {
     const wrapper = render({
       useCPUSForPeopleResults: true,
       crossProductSearchClient: {
@@ -173,7 +202,7 @@ describe('ConfluenceQuickSearchContainer', () => {
 
           return Promise.resolve(EMPTY_CROSS_PRODUCT_SEARCH_RESPONSE);
         },
-        getAbTestData(scope: Scope, searchSession: SearchSession) {
+        getAbTestData(scope: Scope) {
           return Promise.resolve(DEFAULT_AB_TEST);
         },
       },
@@ -184,6 +213,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults.results.people).toEqual([
@@ -212,6 +242,7 @@ describe('ConfluenceQuickSearchContainer', () => {
       'query',
       sessionId,
       100,
+      0,
     );
 
     expect(searchResults.results.people).toEqual([