+++ packages/media/media-test-helpers/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-test-helpers
 
+## 22.0.0
+- [major] [7c17b35107](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7c17b35107):
+
+  - Updates react and react-dom peer dependencies to react@^16.8.0 and react-dom@^16.8.0. To use this package, please ensure you use at least this version of react and react-dom.
+
 ## 21.4.0
 - [minor] [6f463c4d88](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6f463c4d88):